extract sidebar
